requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.1
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.60.1
idna==3.6
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
pypng==0.20220715.0
python-dotenv==1.0.1
python3-openid==3.2.0
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
